@@ -2,7 +2,7 @@ package com.mao.test;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 import com.mao.dev.R;
