@@ -2,13 +2,11 @@ package com.master.cvtutorial;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
-
-import com.master.cvtutorial.R;
 
 public class MainActivity extends AppCompatActivity {
 
